Homework completed. ~ 1 hour.
Ref. to Dictionaries slides to setup melons.py as a dict w/in dict for easy assess

Extra Q from this assignment:
Why is the dictionary more effective/flexible this way?
B/c it allows fellow users to easily understand which factors are being called.
For example, items[name][seedless] and items[name][prices] calls for the seedless and price keys within the outter dicts.